% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_plot_fnct.R
\name{adv_tree_plot}
\alias{adv_tree_plot}
\title{Plot Taxonomy Tree Graph}
\usage{
adv_tree_plot(edgelist, usage_file, out_file, roots, out_width = 5760,
  out_height = 5760, title)
}
\arguments{
\item{edgelist}{A csv consisting of two columns (parent and children) of
relationships between elements}

\item{usage_file}{An optional csv of usage counts for the elements with two
columns, the element and the count for the given time period}

\item{out_file}{The output file for the png image}

\item{roots}{An optional vector of element names that should act as roots.
Their label text size will be increased relative to other elements}

\item{out_width}{the size of the output width in pixels}

\item{out_height}{output height in pixels}

\item{title}{title of the plot}
}
\value{
A png image of the graph
}
\description{
Saves a png image of a network tree (from an igraph object) at a user
specified size with the given plot tile.
}
