% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_animation.R
\name{check_child_path}
\alias{check_child_path}
\title{Recursively Check Children}
\usage{
check_child_path(g_old, g_new, v)
}
\arguments{
\item{g_old}{an igraph graph, the working graph}

\item{g_new}{an igraph graph, the desired graph}

\item{v}{an igraph vertex name present in the old graph, the end node in the
path}
}
\value{
boolean; True if all parents are esired and in correct order, False
otherwise
}
\description{
Method for checking child path for desired node.
}
