% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_animation.R
\name{children_desired}
\alias{children_desired}
\title{Should Children be Kept?}
\usage{
children_desired(g_old, g_new, v)
}
\arguments{
\item{g_old}{igraph graph, the working graph}

\item{g_new}{igraph graph, the desired graph}

\item{v}{an igraph vertex name whose children will be evaluated}
}
\value{
boolean, True if at least one child is in the desired graph, False
  otherwise
}
\description{
Determines if the children of a vertex are desired or not.
}
