% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{choose_linkbase}
\alias{choose_linkbase}
\title{Choose Linkbase}
\usage{
choose_linkbase(linkbase)
}
\arguments{
\item{linkbase}{Linkbase to choose, either Calculation of Presentation}
}
\value{
Integer value of the excel sheet that needs to be read in
}
\description{
A method for choosing the correct Calculation or Presentation linkbase sheet
for importing taxonomy relationships.
}
\examples{
choose_linkbase("Calculation")
}
