% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_animation.R
\name{create_children}
\alias{create_children}
\title{Create New Child Vertex}
\usage{
create_children(g_old, g_new, v, filename)
}
\arguments{
\item{g_old}{an igraph graph, the working graph}

\item{g_new}{an igraph graph, the desired graph}

\item{v}{an igraph vertex name present in the working graph, whose children will
be created}

\item{filename}{The output filename}
}
\value{
an igraph graph, with the correct children nodes in place.
}
\description{
Creates a new child vertex, working recursively
}
