% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{create_graph}
\alias{create_graph}
\title{Create Network Graph}
\usage{
create_graph(statement_of_interest = statement, root_nodes = NA)
}
\arguments{
\item{statement_of_interest}{A dataframe of the statement of interest, often
generated from a call to find_statement. Must have the column names parent
and child.}

\item{root_nodes}{A character vector of element names to place at the highest
level of the graph. This function increases the root_nodes display size.}
}
\value{
An igraph object from the list of edges in the statement dataframe
}
\description{
A function for creating a network graph from a statement of interest in the
form of a dataframe.
}
