% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_clean.R
\name{create_graph_list}
\alias{create_graph_list}
\title{List of graphs}
\usage{
create_graph_list(statement_names, df, link)
}
\arguments{
\item{statement_names}{A vector of statement names}

\item{df}{A dataframe of the link (e.g. calc_link_dirty)}

\item{link}{The linkbase to use, either Presentation or Calculation}
}
\value{
A list of igraph objects
}
\description{
Creates a list of igraph objects which can be used to iterate over
}
