% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{create_statement_tree}
\alias{create_statement_tree}
\title{Create Statement Tree}
\usage{
create_statement_tree(taxonomy_path, linkbase, statement, title, image_path)
}
\arguments{
\item{taxonomy_path}{Path to xls/xlsx file}

\item{linkbase}{Linkbase to choose, either Calculation of Presentation}

\item{statement}{Title of statement of interest as string. Incorrect or vague
values will throw error. Ideally the statement argument would be in the
format on
http://xbrlview.fasb.org/yeti/resources/yeti-gwt/Yeti.jsp#tax~(id~161*v~5016)!net~(a~3191*l~776)!lang~(code~en-us)!rg~(rg~32*p~12)
 like like '124000 - Statement - Statement of Income.}

\item{title}{Title of plot}

\item{image_path}{Output filename}
}
\value{
Plot, in png format
}
\description{
Creates statement tree from scratch excel file
}
