% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_clean.R
\name{edge_to_df}
\alias{edge_to_df}
\title{Read and Convert to Dataframe}
\usage{
edge_to_df(edge_file, sheet = 1, parent_col = 1, child_col = 2,
  fill = TRUE, column_header = TRUE)
}
\arguments{
\item{edge_file}{The path to an excel or csv edgefile that contains at
minimum two columns, a parent and a child.}

\item{sheet}{If an excel document, the sheet that contains the edges}

\item{parent_col}{The column that contains the parent element names}

\item{child_col}{The column that contains the child element names}

\item{fill}{Do parent columns need to be filed?}

\item{column_header}{do column headers exist?
_________________________#
| parent_1 | child_1 |
|          | child_2 |
|          | child_3 |
| parent_2 | child_1 |
|          | child_2 |}
}
\value{
A tibble that con be converted to a dataframe and saved as an Rdata
file
}
\description{
Reads a file in as an excel or a csv and creates an edgelist
}
