% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{excel_to_df}
\alias{excel_to_df}
\title{Excel Sheet to dataframe}
\usage{
excel_to_df(path, linkbase)
}
\arguments{
\item{path}{Path to xls/xlsx file}

\item{linkbase}{Linkbase to choose, either Calculation of Presentation}
}
\value{
dataframe of excel sheet
}
\description{
A method for extracting the correct linkbase from excel sheet and reading in
as a dataframe.
}
