% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{extract_subgraph}
\alias{extract_subgraph}
\title{Create Subgraph from Root Node}
\usage{
extract_subgraph(graph = g, element_of_interest)
}
\arguments{
\item{graph}{The parent graph from which to extract a subgraph}

\item{element_of_interest}{Root node of returned graph as a string. (e.g.
"Revenues")}
}
\value{
A subgraph with root \code{element_of_interest}
}
\description{
Extracts a subgraph from a larger graph by grabbing children elements
(recursively) of an element of interest.
}
