% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{find_statement}
\alias{find_statement}
\title{Extract Financial Statement of Interest}
\usage{
find_statement(statement = "Statement of Income (Including Gross Margin)",
  taxonomy, linkbase = "Calculation")
}
\arguments{
\item{statement}{Title of statement of interest as string. Incorrect or vague
values will throw error. Ideally the statement argument would be in the
format on
http://xbrlview.fasb.org/yeti/resources/yeti-gwt/Yeti.jsp#tax~(id~161*v~5016)!net~(a~3191*l~776)!lang~(code~en-us)!rg~(rg~32*p~12)
 like like '124000 - Statement - Statement of Income.}

\item{taxonomy}{dataframe of the taxonomy of interest}

\item{linkbase}{The linkbase to choose, either Calculation or Presentation}
}
\value{
The statement of interest with each row as an element, the
  corresponding parent, the element label, depth, and if calculation linkbase
  the weight
}
\description{
A function for extracting the financial statement of interest from the given
taxonomy.
}
