% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_clean.R
\name{get_stmt_names}
\alias{get_stmt_names}
\title{Get Statement Names}
\usage{
get_stmt_names(df, link)
}
\arguments{
\item{df}{The dataframe of the link containing elements, definitions, and
linkroles (e.g. calc_link_dirty)}

\item{link}{The linkbase to use, either Presentation or Calculation}
}
\value{
A vector of statement names
}
\description{
Functionality for obtaining statement names from a given taxonomy, either
presentation link, or calculation
}
