% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_panel_fncts.R
\name{panel_trees}
\alias{panel_trees}
\title{Plot Multiple Graphs}
\usage{
panel_trees(graphs, out_file, main_title, graph_titles, rows, cols)
}
\arguments{
\item{graphs}{a list or vector of graphs to plot}

\item{out_file}{the output file to save to}

\item{main_title}{the overall title of the graph}

\item{graph_titles}{a parallel vector of titles whose indices correspond to}

\item{rows}{The number of rows}

\item{cols}{The number of columns
the graph vector.}
}
\description{
Method for plotting panels of statements
}
