% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot Network Graph}
\usage{
plot_graph(g, filename, title)
}
\arguments{
\item{g}{Graph to visualize}

\item{filename}{Output filename}

\item{title}{Title of plot}
}
\value{
The plot, in png format
}
\description{
Saves the network graph in a tree layout as a png of a large size (5760x5760
pixels)
}
