% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_panel_fncts.R
\name{plot_single_graph}
\alias{plot_single_graph}
\title{Plot Graph}
\usage{
plot_single_graph(g, title, title_size = 5, display_names = FALSE)
}
\arguments{
\item{g}{The graph to plot}

\item{title}{The title of the graph}

\item{title_size}{Size of title}

\item{display_names}{Whether or not to display names}
}
\value{
A graph image
}
\description{
Plots a single tree graph
}
