% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_animation.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot Tree}
\usage{
plot_tree(g, outname)
}
\description{
Need to create the equivalent of a static variable so that whenever this
function is run, it increments the counter by one.  That way filenames
will be sequential
}
\details{
Function for plotting tree with certain color and plotting parameters

@param g The graph to plot.
 @param outname Name of the output file.
}
