% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_animation.R
\name{transform_tree}
\alias{transform_tree}
\title{Transform Tree}
\usage{
transform_tree(g_old, g_new, file)
}
\arguments{
\item{g_old}{The old/working igraph}

\item{g_new}{The desired igraph}

\item{file}{Image output name}
}
\value{
An igraph object identical to g_new. Creates images and saves them
to the respective location.
}
\description{
Entry function for transforming a taxonomy tree from an "old" status to a
"new status".
}
