% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_animation.R
\name{vertex_decision}
\alias{vertex_decision}
\title{Retain or Delete Vertex}
\usage{
vertex_decision(g_old, g_new, v, file)
}
\arguments{
\item{g_old}{an igraph graph, the working graph}

\item{g_new}{an igraph graph, the desired graph}

\item{v}{an igraph vertex name present in the "old" graph}

\item{file}{The output filename for the image}
}
\value{
an igraph graph with the appropriate edges deleted
}
\description{
Method for retaining or deleting a vertex given its presence in the desired
graph
}
