% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_animation.R
\name{vertex_parents}
\alias{vertex_parents}
\title{Vertex Parents}
\usage{
vertex_parents(g, v)
}
\arguments{
\item{g}{an igraph graph}

\item{v}{an igraph vertex name}
}
\value{
a list containing the parent vertices (igraph.vs object) and parent
  edges (igraph.es object)
}
\description{
Finds the parents of the current vertex.
}
