% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbrl_taxonomy_functions.R
\name{visualize_differences}
\alias{visualize_differences}
\title{Visualize Differences Between Graphs}
\usage{
visualize_differences(edgelist_large, edgelist_small)
}
\arguments{
\item{edgelist_large}{Larger edgelist}

\item{edgelist_small}{smaller edgelist.  Only those elements that are not in
the smaller edgelist will be highlighted}
}
\value{
returns graph with pinkish nodes for those not in other statement
}
\description{
Visualize the differences between two edgelists
}
