% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_panel_fncts.R
\name{plot_single_graph}
\alias{plot_single_graph}
\title{Plot Graph}
\usage{
plot_single_graph(g, title, title_size = 5, display_names = FALSE,
  xlimit = c(-1, 1), ylimit = c(-1, 1), zoom = 1, legend = FALSE)
}
\arguments{
\item{g}{The graph to plot}

\item{title}{The title of the graph}

\item{title_size}{Size of title}

\item{display_names}{Whether or not to display names}

\item{xlimit}{The xlimit of the plot to zoom to (max is c(-1,1))}

\item{ylimit}{The ylimit of the plot to zoom to (max is c(-1,1))}

\item{zoom}{Optional zoom parameter for vertex and label size}

\item{legend}{True if lengend is desired}
}
\value{
A graph image
}
\description{
Plots a single tree graph
}
